Organize authToken tests into distinct areas
- One describe per file
- Tests for Image Tag
- Tests for URL generation
- Tests for genearting tokens using utils
